[config] move state_sync_config to own file
per title
